% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{gt_as_dt}
\alias{gt_as_dt}
\title{Get a dataframe with lubridate dates for the gtfs stop_times_df}
\usage{
gt_as_dt(stop_times_df)
}
\arguments{
\item{stop_times_df}{a gtfsr$stop_times_df dataframe}
}
\value{
an dataframe with arrival and departure time set to lubridate types
}
\description{
Get a dataframe with lubridate dates for the gtfs stop_times_df
}
\examples{
gt_as_dt(stop_times_df)
stop_times_dt <- stop_times_df_as_dt(some_stops)
#plot the histogram of departure times by hour
hist(hour(stop_times_dt$departure_time))
}
