% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencies.R
\name{route_frequency}
\alias{route_frequency}
\title{Get stop frequency for buses aggregated up to routes}
\usage{
route_frequency(gtfs_obj, start_hour = 6, end_hour = 22, dow = c(1, 1, 1,
  1, 1, 0, 0))
}
\arguments{
\item{gtfs_obj}{a standard gtfsr object}

\item{start_time, }{}

\item{end_time, }{}

\item{service}{e.g. "weekend" or "saturday"}
}
\value{
route_headways a dataframe of route headways
}
\description{
should take:
}
